backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
python-multipart
python-jose[cryptography]
pydantic
azure-storage-blob
